Simplify CLI, remove Clap dependency

+++ ast/src/nodes/literal.rs
@@ -9,9 +9,9 @@ pub struct LiteralNode {
 #[derive(Debug)]
 pub enum LiteralKind {
   FloatDecimal(f64),
-  IntDecimal(i128),
-  IntOctal(i128),
-  IntHex(i128),
-  IntBinary(i128),
+  IntDecimal(i32),
+  IntOctal(i32),
+  IntHex(i32),
+  IntBinary(i32),
   Str(String),
 }
